# ♻️ Refactor: replace isInCharset with bytes.IndexByte