# Description

The standard library `bytes` already includes `bytes.IndexByte`, which is implemented in assembly language. Its performance is superior to ours. Additionally, replacing `isInCharset` with the standard library function can reduce maintenance costs.

```go
func BenchmarkIsInCharset(b *testing.B) {
	t := []byte("hello, world!")
	for i := 0; i < b.N; i++ {
		_ = isInCharset('!', t)
		_ = isInCharset(',', t)
		_ = isInCharset('w', t)
		_ = isInCharset('a', t)
		_ = isInCharset('b', t)
		_ = isInCharset('c', t)
	}
}
```

```go
func BenchmarkIsInCharset(b *testing.B) {
	t := []byte("hello, world!")
	for i := 0; i < b.N; i++ {
		_ = bytes.IndexByte(t, '!') != -1
		_ = bytes.IndexByte(t, ',') != -1
		_ = bytes.IndexByte(t, 'w') != -1
		_ = bytes.IndexByte(t, 'a') != -1
		_ = bytes.IndexByte(t, 'b') != -1
		_ = bytes.IndexByte(t, 'c') != -1
	}
}
```

Benchmark test result

```
goos: linux
goarch: amd64
pkg: github.com/gofiber/fiber/v3
cpu: AMD EPYC 7763 64-Core Processor                
              │   ori.txt   │               std.txt               │
              │   sec/op    │   sec/op     vs base                │
IsInCharset-4   36.97n ± 3%   19.39n ± 1%  -47.55% (p=0.000 n=25)
```

## Type of change

- [x] Performance improvement (non-breaking change which improves efficiency)
- [x] Code consistency (non-breaking change which improves code reliability and robustness)

## Checklist

Before you submit your pull request, please make sure you meet these requirements:

- [ ] Followed the inspiration of the Express.js framework for new functionalities, making them similar in usage.
- [x] Conducted a self-review of the code and provided comments for complex or critical parts.
- [ ] Updated the documentation in the `/docs/` directory for [Fiber's documentation](https://docs.gofiber.io/).
- [ ] Added or updated unit tests to validate the effectiveness of the changes or new features.
- [x] Ensured that new and existing unit tests pass locally with the changes.
- [ ] Verified that any new dependencies are essential and have been agreed upon by the maintainers/community.
- [x] Aimed for optimal performance with minimal allocations in the new code.
- [ ] Provided benchmarks for the new code to analyze and improve upon.